缺陷[#19296] 【系统管理-配置管理-设备测点配置】有模板的设备，其下只有1个测点时，修改或删除该测点，设备数据被删除
    [修改内容]: 优化新增本体部位测点时设备名称神秘消失的问题
    [描述说明]:
    [问题原因]:
    [解决方法]:
    [测试建议]:

+++ iot-center-platform/src/main/java/com/yjh/platform/module/device/service/TStdDevicemeteService.java
@@ -56,7 +56,7 @@ public class TStdDevicemeteService{
             List<TDictBusiness> list = tDictBusinessDao.select(null,tStdDeviceMeteDetail.getCustomType(),null,null,null,null,null);
             stdDevice.setCustomName(list.get(0).getDictNote());
             tStdDeviceDao.add(stdDevice);
-            if(tStdDevicemeteDao.selectByDevCus(tStdDeviceMeteDetail.getDeviceId(),"101").size()==0){
+            if(tStdDevicemeteDao.selectByDevCus(tStdDeviceMeteDetail.getDeviceId(),"101").size()==0 && !(tStdDeviceMeteDetail.getCustomType().equals("101"))){
                 tStdDeviceDao.deleteByUnionKeys(tStdDeviceMeteDetail.getDeviceId(),"101");  //删除之前本体部位的设备
             }
         }
